@@ -1,13 +1,14 @@
 package nota.model;
 
-import net.minecraft.registry.RegistryKeys;
-import net.minecraft.registry.entry.RegistryEntry;
-import net.minecraft.sound.SoundEvent;
-import net.minecraft.sound.SoundEvents;
+import net.minecraft.core.Holder;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.sounds.SoundEvent;
+import net.minecraft.sounds.SoundEvents;
 import nota.Nota;
 
 import java.util.HashMap;
 import java.util.Map;
+import java.util.concurrent.atomic.AtomicReference;
 
 /**
  * Version independent Spigot sounds.
@@ -52,7 +53,7 @@ public enum Sound {
 	 * @return org.bukkit.Sound enum
 	 */
 	public static SoundEvent getFromBukkitName(String bukkitSoundName) {
-		return NOTE_BANJO.getSound(bukkitSoundName);
+		return getSound(bukkitSoundName);
 	}
 
 	private SoundEvent getSound() {
@@ -68,59 +69,89 @@ public enum Sound {
 		return null;
 	}
 
-	private SoundEvent getSound(String name) {
+	private static SoundEvent getSound(String name) {
 		switch(name) {
 			case "NOTE_PIANO", "BLOCK_NOTE_HARP", "BLOCK_NOTE_BLOCK_HARP" -> {
-				return bullshit(SoundEvents.BLOCK_NOTE_BLOCK_HARP);
+				return bullshit(SoundEvents.NOTE_BLOCK_HARP);
 			}
 			case "NOTE_BASS", "BLOCK_NOTE_BASS", "BLOCK_NOTE_BLOCK_BASS" -> {
-				return bullshit(SoundEvents.BLOCK_NOTE_BLOCK_BASS);
+				return bullshit(SoundEvents.NOTE_BLOCK_BASS);
 			}
 			case "NOTE_BASS_DRUM", "BLOCK_NOTE_BASEDRUM", "BLOCK_NOTE_BLOCK_BASEDRUM" -> {
-				return bullshit(SoundEvents.BLOCK_NOTE_BLOCK_BASEDRUM);
+				return bullshit(SoundEvents.NOTE_BLOCK_BASEDRUM);
 			}
 			case "NOTE_SNARE_DRUM", "BLOCK_NOTE_SNARE", "BLOCK_NOTE_BLOCK_SNARE" -> {
-				return bullshit(SoundEvents.BLOCK_NOTE_BLOCK_SNARE);
+				return bullshit(SoundEvents.NOTE_BLOCK_SNARE);
 			}
 			case "NOTE_STICKS", "BLOCK_NOTE_HAT", "BLOCK_NOTE_BLOCK_HAT" -> {
-				return bullshit(SoundEvents.BLOCK_NOTE_BLOCK_HAT);
+				return bullshit(SoundEvents.NOTE_BLOCK_HAT);
 			}
 			case "NOTE_BASS_GUITAR", "BLOCK_NOTE_GUITAR", "BLOCK_NOTE_BLOCK_GUITAR" -> {
-				return bullshit(SoundEvents.BLOCK_NOTE_BLOCK_GUITAR);
+				return bullshit(SoundEvents.NOTE_BLOCK_GUITAR);
 			}
 			case "NOTE_FLUTE", "BLOCK_NOTE_FLUTE", "BLOCK_NOTE_BLOCK_FLUTE" -> {
-				return bullshit(SoundEvents.BLOCK_NOTE_BLOCK_FLUTE);
+				return bullshit(SoundEvents.NOTE_BLOCK_FLUTE);
 			}
 			case "NOTE_BELL", "BLOCK_NOTE_BELL", "BLOCK_NOTE_BLOCK_BELL" -> {
-				return bullshit(SoundEvents.BLOCK_NOTE_BLOCK_BELL);
+				return bullshit(SoundEvents.NOTE_BLOCK_BELL);
 			}
 			case "NOTE_XYLOPHONE", "BLOCK_NOTE_XYLOPHONE", "BLOCK_NOTE_BLOCK_XYLOPHONE" -> {
-				return bullshit(SoundEvents.BLOCK_NOTE_BLOCK_XYLOPHONE);
+				return bullshit(SoundEvents.NOTE_BLOCK_XYLOPHONE);
 			}
 			case "NOTE_PLING", "BLOCK_NOTE_PLING", "BLOCK_NOTE_BLOCK_PLING" -> {
-				return bullshit(SoundEvents.BLOCK_NOTE_BLOCK_PLING);
+				return bullshit(SoundEvents.NOTE_BLOCK_PLING);
 			}
 			case "BLOCK_NOTE_BLOCK_IRON_XYLOPHONE" -> {
-				return bullshit(SoundEvents.BLOCK_NOTE_BLOCK_IRON_XYLOPHONE);
+				return bullshit(SoundEvents.NOTE_BLOCK_IRON_XYLOPHONE);
 			}
 			case "BLOCK_NOTE_BLOCK_COW_BELL" -> {
-				return bullshit(SoundEvents.BLOCK_NOTE_BLOCK_COW_BELL);
+				return bullshit(SoundEvents.NOTE_BLOCK_COW_BELL);
 			}
 			case "BLOCK_NOTE_BLOCK_DIDGERIDOO" -> {
-				return bullshit(SoundEvents.BLOCK_NOTE_BLOCK_DIDGERIDOO);
+				return bullshit(SoundEvents.NOTE_BLOCK_DIDGERIDOO);
 			}
 			case "BLOCK_NOTE_BLOCK_BIT" -> {
-				return bullshit(SoundEvents.BLOCK_NOTE_BLOCK_BIT);
+				return bullshit(SoundEvents.NOTE_BLOCK_BIT);
 			}
 			case "BLOCK_NOTE_BLOCK_BANJO" -> {
-				return bullshit(SoundEvents.BLOCK_NOTE_BLOCK_BANJO);
+				return bullshit(SoundEvents.NOTE_BLOCK_BANJO);
 			}
 		}
-		return bullshit(SoundEvents.BLOCK_NOTE_BLOCK_BASS);
+		return bullshit(SoundEvents.NOTE_BLOCK_BASS);
 	}
 
-	private SoundEvent bullshit(RegistryEntry.Reference<SoundEvent> reference) {
-		return Nota.getAPI().getServer().getRegistryManager().get(RegistryKeys.SOUND_EVENT).get(reference.registryKey());
+//	private SoundEvent bullshit(Holder.Reference<SoundEvent> reference) {
+//		MinecraftServer server = Nota.getAPI().getServer();
+//		RegistryAccess registryAccess = server.registryAccess();
+//		Registry<SoundEvent> soundEventRegistry = null;
+//
+//		Registry<SoundEvent>[] result = new Registry[1];
+//
+//		registryAccess.registries().forEach(entry -> {
+//			if (entry.key().equals(Registries.SOUND_EVENT)) {
+//				result[0] = (Registry<SoundEvent>) entry.value();
+//			}
+//		});
+//
+//		if (result[0] == null) {
+//			throw new IllegalStateException("SOUND_EVENT registry not found!");
+//		}
+//
+//		Optional<Holder.Reference<SoundEvent>> soundEvent = result[0].get(reference.key());
+//
+//		// weird java magic
+//		return soundEvent.map(Holder.Reference::value).orElseGet(SoundEvents.NOTE_BLOCK_HARP::value);
+//	}
+
+	private static SoundEvent bullshit(Holder.Reference<SoundEvent> reference) {
+		AtomicReference<SoundEvent> result = new AtomicReference<>();
+
+		Nota.getAPI().getServer().registryAccess().registries().forEach(entry -> {
+			if (entry.key() == Registries.SOUND_EVENT)
+				result.set((SoundEvent) entry.value().get(reference.key().location()).get().value());
+		});
+
+		return result.get();
 	}
 
 	/**
