@@ -1,7 +1,6 @@
 package nota.player;
 
-import net.minecraft.entity.player.PlayerEntity;
-
+import net.minecraft.server.level.ServerPlayer;
 import nota.model.Playlist;
 import nota.model.Song;
 
@@ -38,6 +37,6 @@ public abstract class RangeSongPlayer extends SongPlayer {
 	 * @param player in range
 	 * @return ability to hear the current RangeSongPlayer
 	 */
-	public abstract boolean isInRange(PlayerEntity player);
+	public abstract boolean isInRange(ServerPlayer player);
 
 }
